refactor: move indentContent to FormatterUtil
