+++ src/util/formatter_util.ts
@@ -381,4 +381,49 @@ export default class FormatterUtil {
 
     return output.trimEnd();
   }
+
+  /**
+   * Indents content and splits content across multiple lines.
+   * @param content the content to indent.
+   * @param indentSize the size of the indent.
+   */
+  public static indentContent(content: string, indentSize: number): string {
+    const contentLines = content.split("\n");
+    const resolvedLines: string[] = [];
+
+    const outputElement = TerminalUtil.getOutputElement();
+    const charactersPerLine = this.getCharactersPerLine(outputElement);
+
+    for (const line of contentLines) {
+      if (line.length + indentSize >= charactersPerLine) {
+        const chunks = this.toChunks(line, charactersPerLine - indentSize);
+        resolvedLines.push(...chunks);
+      } else {
+        resolvedLines.push(line);
+      }
+    }
+
+    const indent = " ".repeat(indentSize);
+    const lines = resolvedLines.map((line) => `${indent}${line}`);
+    return lines.join("\n");
+  }
+
+  /**
+   * Converts the `str` into chunks with a maximum size of `size`.
+   *
+   * @param str the string to convert.
+   * @param size the size of the chunks.
+   */
+  private static toChunks(str: string, size: number): string[] {
+    const numChunks = Math.ceil(str.length / size);
+    const chunks = new Array(numChunks);
+
+    let start = 0;
+    for (let i = 0; i < numChunks; ++i) {
+      chunks[i] = str.substring(start, start + size);
+      start += size;
+    }
+
+    return chunks;
+  }
 }
